Phase 2: per-page JS entries replace the app.js monolith
Every page previously loaded app.js, which statically imported 30 modules
(~1.8 MB) and dispatched on getElementById guards. Now each page loads a
thin entry from assets/js/entries/ that imports only what it uses:

- assets/js/core-boot.js: the shared boot extracted from app.js — theme,
  PWA, storage migrations, Bible metadata, tooltip/modal/ref wiring,
  version picker, sidebar Search, history widget, deferred term/place
  tagging, and the window.BibleUI API. boot(pageInit, {early}) preserves
  app.js's init ordering (reader ran before autoTagRefs; the rest after).
- 15 entries: generic (57 pages) + home, reader (read/ and settings/read/),
  memorize, biblepedia, timeline, church-history, maps, timelapse,
  wordcloud, lib-reader, lib-browser, lib-progress, apocrypha, settings.
- Base.astro: new entry prop (default "generic") replaces the app.js tag;
  the 14 feature pages declare theirs. Timelapse's inline app.js?v=2 tag
  and the verbatim statics (topics/index, word/index, 4 _template
  scaffolds) now point at their entries. Dead window.BibleUI.init() relic
  removed from compare/ (classic script ran before modules — never fired).
- sw.js: precache lists core-boot + entries instead of app.js; adds
  synoptic.js and biblepedia.js (pre-existing precache gaps); APP_CACHE_V
  v184 → v185.

Measured transitive payload: generic pages 1,800 KB → 509 KB (−72%);
reader 779 KB; maps 694 KB. All 74 pages build; node --check passes on
every JS file; dist/ shows one entry per page and zero app.js references.
