+++ assets/js/entries/biblepedia.js
@@ -1,0 +1,7 @@
+/* entries/biblepedia.js — Biblepedia encyclopedia entry (biblepedia/). */
+import { boot } from '../core-boot.js';
+import { initBiblepediaPage } from '../biblepedia.js';
+
+boot(function () {
+  initBiblepediaPage();
+});
