+++ assets/js/entries/church-history.js
@@ -1,0 +1,7 @@
+/* entries/church-history.js — Church-history timeline entry (church-history/). */
+import { boot } from '../core-boot.js';
+import { initChurchTimelinePage } from '../timeline.js';
+
+boot(function () {
+  initChurchTimelinePage();
+});
