+++ assets/js/entries/lib-progress.js
@@ -1,0 +1,7 @@
+/* entries/lib-progress.js — Library reading-history entry (library/progress/). */
+import { boot } from '../core-boot.js';
+import { initLibProgressPage } from '../lib-progress.js';
+
+boot(function () {
+  initLibProgressPage();
+});
