+++ assets/js/entries/maps.js
@@ -1,0 +1,7 @@
+/* entries/maps.js — Bible maps entry (maps/). */
+import { boot } from '../core-boot.js';
+import { initMapsPage } from '../maps.js';
+
+boot(function () {
+  initMapsPage();
+});
