+++ assets/js/entries/memorize.js
@@ -1,0 +1,7 @@
+/* entries/memorize.js — Memorisation entry (discipline/ hosts #mem-list). */
+import { boot } from '../core-boot.js';
+import { initMemorizePage } from '../daily.js';
+
+boot(function () {
+  initMemorizePage();
+});
