+++ assets/js/entries/timeline.js
@@ -1,0 +1,7 @@
+/* entries/timeline.js — Biblical timeline entry (timeline/). */
+import { boot } from '../core-boot.js';
+import { initTimelinePage } from '../timeline.js';
+
+boot(function () {
+  initTimelinePage();
+});
