+++ assets/js/entries/wordcloud.js
@@ -1,0 +1,7 @@
+/* entries/wordcloud.js — Word-cloud entry (wordcloud/). */
+import { boot } from '../core-boot.js';
+import { initWordCloudPage } from '../wordcloud.js';
+
+boot(function () {
+  initWordCloudPage();
+});
