+++ src/pages/apocrypha/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Apocrypha — Kingdom Bible Study"} description={"Read complete Catholic, Orthodox, and LXX Bible versions including the full deuterocanonical books — Douay-Rheims, WEB-CE, KJV Apocrypha, Brenton's LXX."} bodyAttrs={{"class":"reader-page"}}>
+<Base entry="apocrypha" title={"Apocrypha — Kingdom Bible Study"} description={"Read complete Catholic, Orthodox, and LXX Bible versions including the full deuterocanonical books — Douay-Rheims, WEB-CE, KJV Apocrypha, Brenton's LXX."} bodyAttrs={{"class":"reader-page"}}>
 <Fragment slot="head">
 <meta property="og:type" content="website"><meta property="og:title" content="Apocrypha — Kingdom Bible Study"><meta property="og:description" content="Complete Catholic, Orthodox, and LXX Bible editions including Tobit, Judith, Sirach, Wisdom, Maccabees, and all deuterocanonical books."><meta name="twitter:card" content="summary"><meta name="twitter:title" content="Apocrypha — Kingdom Bible Study"><meta name="twitter:description" content="Full Catholic, Orthodox, and LXX Bible — Douay-Rheims, WEB-CE, KJV Apocrypha, Brenton's LXX.">
 </Fragment>
