+++ src/pages/biblepedia/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Biblepedia — Kingdom Bible Study"} description={"Biblepedia — a wiki-style Bible reference with 4,500+ articles on Biblical people, places, concepts, and topics."} footer={false}>
+<Base entry="biblepedia" title={"Biblepedia — Kingdom Bible Study"} description={"Biblepedia — a wiki-style Bible reference with 4,500+ articles on Biblical people, places, concepts, and topics."} footer={false}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../assets/css/dictionary.css"><link rel="stylesheet" href="../assets/css/biblepedia.css">
 </Fragment>
