+++ src/pages/church-history/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Church History Timeline — Kingdom Bible Study"} description={"Interactive church history timeline from Pentecost to the present — the Apostolic Age, the Fathers, the Councils, the Reformation, and beyond."} footer={false}>
+<Base entry="church-history" title={"Church History Timeline — Kingdom Bible Study"} description={"Interactive church history timeline from Pentecost to the present — the Apostolic Age, the Fathers, the Councils, the Reformation, and beyond."} footer={false}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../assets/css/timeline.css">
 </Fragment>
