+++ src/pages/compare/index.astro
@@ -449,8 +449,6 @@ import Base from '../../layouts/Base.astro';
         navigate(val);
       });
 
-      // BibleUI for tooltip/modal system
-      if (window.BibleUI) window.BibleUI.init();
     })();
   </script>
 </Fragment>
