+++ src/pages/discipline/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Discipline — Kingdom Bible Study"} description={"Daily spiritual disciplines — reading plans, devotionals, memory, journal, and worship notes."}>
+<Base entry="memorize" title={"Discipline — Kingdom Bible Study"} description={"Daily spiritual disciplines — reading plans, devotionals, memory, journal, and worship notes."}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../assets/css/memorize.css"><link rel="stylesheet" href="../assets/css/devotionals.css"><link rel="stylesheet" href="../assets/css/discipline.css">
 </Fragment>
