+++ src/pages/index.astro
@@ -2,7 +2,7 @@
 import Base from '../layouts/Base.astro';
 ---
 
-<Base title={"Kingdom Bible Study — Daily"} description={"Daily Bible reading, verse of the day, and devotional."}>
+<Base entry="home" title={"Kingdom Bible Study — Daily"} description={"Daily Bible reading, verse of the day, and devotional."}>
 <Fragment slot="head">
 <meta property="og:type" content="website"><meta property="og:title" content="Kingdom Bible Study — Daily"><meta property="og:description" content="Daily Bible reading, verse of the day, and devotional."><meta name="twitter:card" content="summary"><meta name="twitter:title" content="Kingdom Bible Study — Daily"><meta name="twitter:description" content="Daily Bible reading, verse of the day, and devotional.">
 </Fragment>
