+++ src/pages/library/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Library — Kingdom Bible Study"} description={"Christian library — browse creeds, confessions, catechisms, and church fathers with filtering and an inline reader."} bodyAttrs={{"class":"lb-browser-page"}}>
+<Base entry="lib-browser" title={"Library — Kingdom Bible Study"} description={"Christian library — browse creeds, confessions, catechisms, and church fathers with filtering and an inline reader."} bodyAttrs={{"class":"lb-browser-page"}}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../assets/css/library.css"><link rel="stylesheet" href="../assets/css/lib-browser.css">
 </Fragment>
