+++ src/pages/library/progress/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../../layouts/Base.astro';
 ---
 
-<Base title={"Reading History — Library — Kingdom Bible Study"} description={"Library reading history and tradition coverage."}>
+<Base entry="lib-progress" title={"Reading History — Library — Kingdom Bible Study"} description={"Library reading history and tradition coverage."}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../../assets/css/library.css"><link rel="stylesheet" href="../../assets/css/lib-reader.css"><link rel="stylesheet" href="../../assets/css/lib-progress.css">
 </Fragment>
