+++ src/pages/library/read/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../../layouts/Base.astro';
 ---
 
-<Base title={"Library — Kingdom Bible Study"} description={"Read historic Christian documents — creeds, confessions, catechisms, and church fathers."}>
+<Base entry="lib-reader" title={"Library — Kingdom Bible Study"} description={"Read historic Christian documents — creeds, confessions, catechisms, and church fathers."}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../../assets/css/library.css"><link rel="stylesheet" href="../../assets/css/lib-reader.css">
 </Fragment>
