+++ src/pages/maps/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Biblical Maps — Kingdom Bible Study"} description={"Biblical geography maps — explore the Holy Land, Exodus route, Paul's journeys, and more."} footer={false}>
+<Base entry="maps" title={"Biblical Maps — Kingdom Bible Study"} description={"Biblical geography maps — explore the Holy Land, Exodus route, Paul's journeys, and more."} footer={false}>
 <Fragment slot="head-pre-css">
 <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" integrity="sha384-sHL9NAb7lN7rfvG5lfHpm643Xkcjzp4jFvuavGOndn6pjVqS6ny56CAt3nsEVT4H" crossorigin="">
 </Fragment>
