+++ src/pages/maps/timelapse/index.astro
@@ -73,5 +73,5 @@ import Base from '../../../layouts/Base.astro';
       </div>
 
     </div>
-  </main><script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js" integrity="sha384-cxOPjt7s7Iz04uaHJceBmS+qpjv2JkIHNVcuOrM+YHwZOmJGBXI00mdUXEq65HTH" crossorigin="" is:inline=""></script><script src="../../assets/js/main.js?v=2" is:inline=""></script><script type="module" src="../../assets/js/app.js?v=2" is:inline=""></script>
+  </main><script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js" integrity="sha384-cxOPjt7s7Iz04uaHJceBmS+qpjv2JkIHNVcuOrM+YHwZOmJGBXI00mdUXEq65HTH" crossorigin="" is:inline=""></script><script src="../../assets/js/main.js?v=2" is:inline=""></script><script type="module" src="../../assets/js/entries/timelapse.js" is:inline=""></script>
 </Base>
