+++ src/pages/read/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"The Holy Bible — Kingdom Bible Study"} description={"Read the Bible by reference with version switching."} bodyAttrs={{"class":"reader-page"}}>
+<Base entry="reader" title={"The Holy Bible — Kingdom Bible Study"} description={"Read the Bible by reference with version switching."} bodyAttrs={{"class":"reader-page"}}>
 <Fragment slot="head">
 <meta property="og:type" content="website"><meta property="og:title" id="bsw-og-title" content="The Holy Bible — Kingdom Bible Study"><meta property="og:description" id="bsw-og-desc" content="Read the Bible by reference with version switching."><meta name="twitter:card" content="summary"><meta name="twitter:title" id="bsw-tw-title" content="The Holy Bible — Kingdom Bible Study"><meta name="twitter:description" id="bsw-tw-desc" content="Read the Bible by reference with version switching.">
 </Fragment>
