+++ src/pages/settings/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Settings — Kingdom Bible Study"} description={"Settings — manage offline data and download Bible study features for use without a connection."} footer={false}>
+<Base entry="settings" title={"Settings — Kingdom Bible Study"} description={"Settings — manage offline data and download Bible study features for use without a connection."} footer={false}>
 <Fragment slot="head-end">
 <style is:inline="">
     .set-page { max-width: 820px; margin: 0 auto; padding-bottom: 5rem; }
