+++ src/pages/timeline/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Biblical Timeline — Kingdom Bible Study"} description={"Interactive biblical timeline from Creation to the Consummation — explore every era of redemptive history."} footer={false}>
+<Base entry="timeline" title={"Biblical Timeline — Kingdom Bible Study"} description={"Interactive biblical timeline from Creation to the Consummation — explore every era of redemptive history."} footer={false}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../assets/css/timeline.css">
 </Fragment>
