+++ src/pages/wordcloud/index.astro
@@ -2,7 +2,7 @@
 import Base from '../../layouts/Base.astro';
 ---
 
-<Base title={"Word Cloud — Kingdom Bible Study"} description={"Biblical word cloud — explore the most theologically significant words in Scripture, sized by frequency across the whole Bible or any section."} footer={false}>
+<Base entry="wordcloud" title={"Word Cloud — Kingdom Bible Study"} description={"Biblical word cloud — explore the most theologically significant words in Scripture, sized by frequency across the whole Bible or any section."} footer={false}>
 <Fragment slot="head-end">
 <link rel="stylesheet" href="../assets/css/wordcloud.css">
 </Fragment>
