+++ study-guides/_template/index.html
@@ -159,6 +159,6 @@
     }());
   </script>
 
-  <script type="module" src="../../assets/js/app.js"></script>
+  <script type="module" src="../../assets/js/entries/generic.js"></script>
 </body>
 </html>
