+++ topics/_template-book/deep-dive.html
@@ -132,6 +132,6 @@
   </footer>
 
   <script src="../../assets/js/main.js"></script>
-  <script type="module" src="../../assets/js/app.js"></script>
+  <script type="module" src="../../assets/js/entries/generic.js"></script>
 </body>
 </html>
