+++ topics/_template/index.html
@@ -169,6 +169,6 @@
   </footer>
 
   <script src="../../assets/js/main.js"></script>
-  <script type="module" src="../../assets/js/app.js"></script>
+  <script type="module" src="../../assets/js/entries/generic.js"></script>
 </body>
 </html>
