+++ topics/index.html
@@ -85,6 +85,6 @@
   </footer>
 
     <script src="../assets/js/main.js"></script>
-    <script type="module" src="../assets/js/app.js"></script>
+    <script type="module" src="../assets/js/entries/generic.js"></script>
 </body>
 </html>
