+++ word/index.html
@@ -64,6 +64,6 @@
   </footer>
 
   <script src="../assets/js/main.js"></script>
-  <script type="module" src="../assets/js/app.js"></script>
+  <script type="module" src="../assets/js/entries/generic.js"></script>
 </body>
 </html>
